:recycle: prompt 컴포넌트 분리 및 위치 이동 #57
