+++ index.html
@@ -1,14 +1,13 @@
 <!doctype html>
 <html lang="en">
 <head>
-  <meta charset="UTF-8"/>
-  <link rel="icon" type="image/svg+xml" href="/favicon.svg"/>
-  <meta name="viewport" content="width=device-width, viewport-fit=cover, minimal-ui"/>
+  <meta charset="UTF-8" />
+  <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
+  <meta name="viewport" content="width=device-width, viewport-fit=cover, minimal-ui" />
   <title>NoColored</title>
 </head>
 <body>
 <div id="modal"></div>
-<div id="prompt"></div>
 <div id="root"></div>
 <script type="module" src="/src/main.tsx"></script>
 </body>
