+++ src/app/index.tsx
@@ -2,6 +2,8 @@ import { RouterProvider } from 'react-router-dom';
 
 import router from './router';
 
+import FullScreenPrompt from '@/components/prompt';
+
 import { useSoundSetting } from '@/features/sound';
 
 const App = () => {
@@ -9,6 +11,7 @@ const App = () => {
 
   return (
     <div onClick={playBgm} role='presentation'>
+      <FullScreenPrompt />
       <RouterProvider router={router} />
     </div>
   );
